Moving more functions into individual modules.  Updating utils references.
