+++ src/saltext/vmware_vsphere/modules/vmware_service.py
@@ -112,7 +112,7 @@ def service_start(
         salt '*' vsphere.service_start my.vcenter.location root bad-password 'ntpd' \
         host_names='[esxi-1.host.com, esxi-2.host.com]'
     """
-    service_instance = salt.utils.vmware.get_service_instance(
+    service_instance = saltext.vmware.utils.vmware.get_service_instance(
         host=host,
         username=username,
         password=password,
@@ -254,7 +254,7 @@ def service_stop(
         salt '*' vsphere.service_stop my.vcenter.location root bad-password 'ssh' \
         host_names='[esxi-1.host.com, esxi-2.host.com]'
     """
-    service_instance = salt.utils.vmware.get_service_instance(
+    service_instance = saltext.vmware.utils.vmware.get_service_instance(
         host=host,
         username=username,
         password=password,
@@ -394,7 +394,7 @@ def service_restart(
         salt '*' vsphere.service_restart my.vcenter.location root bad-password 'ntpd' \
         host_names='[esxi-1.host.com, esxi-2.host.com]'
     """
-    service_instance = salt.utils.vmware.get_service_instance(
+    service_instance = saltext.vmware.utils.vmware.get_service_instance(
         host=host,
         username=username,
         password=password,
@@ -540,7 +540,7 @@ def set_service_policy(
         salt '*' vsphere.set_service_policy my.vcenter.location root bad-password 'ntpd' 'automatic' \
         host_names='[esxi-1.host.com, esxi-2.host.com]'
     """
-    service_instance = salt.utils.vmware.get_service_instance(
+    service_instance = saltext.vmware.utils.vmware.get_service_instance(
         host=host,
         username=username,
         password=password,
@@ -628,4 +628,285 @@ def set_service_policy(
 
     return ret
 
+@depends(HAS_PYVMOMI)
+@ignores_kwargs("credstore")
+def get_service_policy(
+    host,
+    username,
+    password,
+    service_name,
+    protocol=None,
+    port=None,
+    host_names=None,
+    verify_ssl=True,
+):
+    """
+    Get the service name's policy for a given host or list of hosts.
 
+    host
+        The location of the host.
+
+    username
+        The username used to login to the host, such as ``root``.
+
+    password
+        The password used to login to the host.
+
+    service_name
+        The name of the service for which to retrieve the policy. Supported service names are:
+          - DCUI
+          - TSM
+          - SSH
+          - lbtd
+          - lsassd
+          - lwiod
+          - netlogond
+          - ntpd
+          - sfcbd-watchdog
+          - snmpd
+          - vprobed
+          - vpxa
+          - xorg
+
+    protocol
+        Optionally set to alternate protocol if the host is not using the default
+        protocol. Default protocol is ``https``.
+
+    port
+        Optionally set to alternate port if the host is not using the default
+        port. Default port is ``443``.
+
+    host_names
+        List of ESXi host names. When the host, username, and password credentials
+        are provided for a vCenter Server, the host_names argument is required to tell
+        vCenter the hosts for which to get service policy information.
+
+        If host_names is not provided, the service policy information will be retrieved
+        for the ``host`` location instead. This is useful for when service instance
+        connection information is used for a single ESXi host.
+
+    verify_ssl
+        Verify the SSL certificate. Default: True
+
+    CLI Example:
+
+    .. code-block:: bash
+
+        # Used for single ESXi host connection information
+        salt '*' vsphere.get_service_policy my.esxi.host root bad-password 'ssh'
+
+        # Used for connecting to a vCenter Server
+        salt '*' vsphere.get_service_policy my.vcenter.location root bad-password 'ntpd' \
+        host_names='[esxi-1.host.com, esxi-2.host.com]'
+    """
+    service_instance = salt.utils.vmware.get_service_instance(
+        host=host,
+        username=username,
+        password=password,
+        protocol=protocol,
+        port=port,
+        verify_ssl=verify_ssl,
+    )
+    valid_services = [
+        "DCUI",
+        "TSM",
+        "SSH",
+        "ssh",
+        "lbtd",
+        "lsassd",
+        "lwiod",
+        "netlogond",
+        "ntpd",
+        "sfcbd-watchdog",
+        "snmpd",
+        "vprobed",
+        "vpxa",
+        "xorg",
+    ]
+    host_names = _check_hosts(service_instance, host, host_names)
+
+    ret = {}
+    for host_name in host_names:
+        # Check if the service_name provided is a valid one.
+        # If we don't have a valid service, return. The service will be invalid for all hosts.
+        if service_name not in valid_services:
+            ret.update(
+                {
+                    host_name: {
+                        "Error": "{} is not a valid service name.".format(service_name)
+                    }
+                }
+            )
+            return ret
+
+        host_ref = _get_host_ref(service_instance, host, host_name=host_name)
+        services = host_ref.configManager.serviceSystem.serviceInfo.service
+
+        # Don't require users to know that VMware lists the ssh service as TSM-SSH
+        if service_name == "SSH" or service_name == "ssh":
+            temp_service_name = "TSM-SSH"
+        else:
+            temp_service_name = service_name
+
+        # Loop through services until we find a matching name
+        for service in services:
+            if service.key == temp_service_name:
+                ret.update({host_name: {service_name: service.policy}})
+                # We've found a match - break out of the loop so we don't overwrite the
+                # Updated host_name value with an error message.
+                break
+            else:
+                msg = "Could not find service '{}' for host '{}'.".format(
+                    service_name, host_name
+                )
+                ret.update({host_name: {"Error": msg}})
+
+        # If we made it this far, something else has gone wrong.
+        if ret.get(host_name) is None:
+            msg = "'vsphere.get_service_policy' failed for host {}.".format(host_name)
+            log.debug(msg)
+            ret.update({host_name: {"Error": msg}})
+
+    return ret
+
+
+@depends(HAS_PYVMOMI)
+@ignores_kwargs("credstore")
+def get_service_running(
+    host,
+    username,
+    password,
+    service_name,
+    protocol=None,
+    port=None,
+    host_names=None,
+    verify_ssl=True,
+):
+    """
+    Get the service name's running state for a given host or list of hosts.
+
+    host
+        The location of the host.
+
+    username
+        The username used to login to the host, such as ``root``.
+
+    password
+        The password used to login to the host.
+
+    service_name
+        The name of the service for which to retrieve the policy. Supported service names are:
+          - DCUI
+          - TSM
+          - SSH
+          - lbtd
+          - lsassd
+          - lwiod
+          - netlogond
+          - ntpd
+          - sfcbd-watchdog
+          - snmpd
+          - vprobed
+          - vpxa
+          - xorg
+
+    protocol
+        Optionally set to alternate protocol if the host is not using the default
+        protocol. Default protocol is ``https``.
+
+    port
+        Optionally set to alternate port if the host is not using the default
+        port. Default port is ``443``.
+
+    host_names
+        List of ESXi host names. When the host, username, and password credentials
+        are provided for a vCenter Server, the host_names argument is required to tell
+        vCenter the hosts for which to get the service's running state.
+
+        If host_names is not provided, the service's running state will be retrieved
+        for the ``host`` location instead. This is useful for when service instance
+        connection information is used for a single ESXi host.
+
+    verify_ssl
+        Verify the SSL certificate. Default: True
+
+    CLI Example:
+
+    .. code-block:: bash
+
+        # Used for single ESXi host connection information
+        salt '*' vsphere.get_service_running my.esxi.host root bad-password 'ssh'
+
+        # Used for connecting to a vCenter Server
+        salt '*' vsphere.get_service_running my.vcenter.location root bad-password 'ntpd' \
+        host_names='[esxi-1.host.com, esxi-2.host.com]'
+    """
+    service_instance = salt.utils.vmware.get_service_instance(
+        host=host,
+        username=username,
+        password=password,
+        protocol=protocol,
+        port=port,
+        verify_ssl=verify_ssl,
+    )
+    valid_services = [
+        "DCUI",
+        "TSM",
+        "SSH",
+        "ssh",
+        "lbtd",
+        "lsassd",
+        "lwiod",
+        "netlogond",
+        "ntpd",
+        "sfcbd-watchdog",
+        "snmpd",
+        "vprobed",
+        "vpxa",
+        "xorg",
+    ]
+    host_names = _check_hosts(service_instance, host, host_names)
+
+    ret = {}
+    for host_name in host_names:
+        # Check if the service_name provided is a valid one.
+        # If we don't have a valid service, return. The service will be invalid for all hosts.
+        if service_name not in valid_services:
+            ret.update(
+                {
+                    host_name: {
+                        "Error": "{} is not a valid service name.".format(service_name)
+                    }
+                }
+            )
+            return ret
+
+        host_ref = _get_host_ref(service_instance, host, host_name=host_name)
+        services = host_ref.configManager.serviceSystem.serviceInfo.service
+
+        # Don't require users to know that VMware lists the ssh service as TSM-SSH
+        if service_name == "SSH" or service_name == "ssh":
+            temp_service_name = "TSM-SSH"
+        else:
+            temp_service_name = service_name
+
+        # Loop through services until we find a matching name
+        for service in services:
+            if service.key == temp_service_name:
+                ret.update({host_name: {service_name: service.running}})
+                # We've found a match - break out of the loop so we don't overwrite the
+                # Updated host_name value with an error message.
+                break
+            else:
+                msg = "Could not find service '{}' for host '{}'.".format(
+                    service_name, host_name
+                )
+                ret.update({host_name: {"Error": msg}})
+
+        # If we made it this far, something else has gone wrong.
+        if ret.get(host_name) is None:
+            msg = "'vsphere.get_service_running' failed for host {}.".format(host_name)
+            log.debug(msg)
+            ret.update({host_name: {"Error": msg}})
+
+    return ret
