+++ src/saltext/vmware_vsphere/modules/vmware_vsphere.py
@@ -132,475 +132,6 @@ def get_proxy_type():
     return "<undefined>"
 
 
-@depends(HAS_ESX_CLI)
-def syslog_service_reload(
-    host, username, password, protocol=None, port=None, esxi_hosts=None, credstore=None
-):
-    """
-    Reload the syslog service so it will pick up any changes.
-
-    host
-        The location of the host.
-
-    username
-        The username used to login to the host, such as ``root``.
-
-    password
-        The password used to login to the host.
-
-    protocol
-        Optionally set to alternate protocol if the host is not using the default
-        protocol. Default protocol is ``https``.
-
-    port
-        Optionally set to alternate port if the host is not using the default
-        port. Default port is ``443``.
-
-    esxi_hosts
-        If ``host`` is a vCenter host, then use esxi_hosts to execute this function
-        on a list of one or more ESXi machines.
-
-    credstore
-        Optionally set to path to the credential store file.
-
-    :return: A standard cmd.run_all dictionary.  This dictionary will at least
-             have a `retcode` key.  If `retcode` is 0 the command was successful.
-
-    CLI Example:
-
-    .. code-block:: bash
-
-        # Used for ESXi host connection information
-        salt '*' vsphere.syslog_service_reload my.esxi.host root bad-password
-
-        # Used for connecting to a vCenter Server
-        salt '*' vsphere.syslog_service_reload my.vcenter.location root bad-password \
-            esxi_hosts='[esxi-1.host.com, esxi-2.host.com]'
-    """
-    cmd = "system syslog reload"
-
-    ret = {}
-    if esxi_hosts:
-        if not isinstance(esxi_hosts, list):
-            raise CommandExecutionError("'esxi_hosts' must be a list.")
-
-        for esxi_host in esxi_hosts:
-            response = salt.utils.vmware.esxcli(
-                host,
-                username,
-                password,
-                cmd,
-                protocol=protocol,
-                port=port,
-                esxi_host=esxi_host,
-                credstore=credstore,
-            )
-            ret.update({esxi_host: response})
-    else:
-        # Handles a single host or a vCenter connection when no esxi_hosts are provided.
-        response = salt.utils.vmware.esxcli(
-            host,
-            username,
-            password,
-            cmd,
-            protocol=protocol,
-            port=port,
-            credstore=credstore,
-        )
-        ret.update({host: response})
-
-    return ret
-
-
-@depends(HAS_ESX_CLI)
-def set_syslog_config(
-    host,
-    username,
-    password,
-    syslog_config,
-    config_value,
-    protocol=None,
-    port=None,
-    firewall=True,
-    reset_service=True,
-    esxi_hosts=None,
-    credstore=None,
-):
-    """
-    Set the specified syslog configuration parameter. By default, this function will
-    reset the syslog service after the configuration is set.
-
-    host
-        ESXi or vCenter host to connect to.
-
-    username
-        User to connect as, usually root.
-
-    password
-        Password to connect with.
-
-    syslog_config
-        Name of parameter to set (corresponds to the command line switch for
-        esxcli without the double dashes (--))
-
-        Valid syslog_config values are ``logdir``, ``loghost``, ``default-rotate`,
-        ``default-size``, ``default-timeout``, and ``logdir-unique``.
-
-    config_value
-        Value for the above parameter. For ``loghost``, URLs or IP addresses to
-        use for logging. Multiple log servers can be specified by listing them,
-        comma-separated, but without spaces before or after commas.
-
-        (reference: https://blogs.vmware.com/vsphere/2012/04/configuring-multiple-syslog-servers-for-esxi-5.html)
-
-    protocol
-        Optionally set to alternate protocol if the host is not using the default
-        protocol. Default protocol is ``https``.
-
-    port
-        Optionally set to alternate port if the host is not using the default
-        port. Default port is ``443``.
-
-    firewall
-        Enable the firewall rule set for syslog. Defaults to ``True``.
-
-    reset_service
-        After a successful parameter set, reset the service. Defaults to ``True``.
-
-    esxi_hosts
-        If ``host`` is a vCenter host, then use esxi_hosts to execute this function
-        on a list of one or more ESXi machines.
-
-    credstore
-        Optionally set to path to the credential store file.
-
-    :return: Dictionary with a top-level key of 'success' which indicates
-             if all the parameters were reset, and individual keys
-             for each parameter indicating which succeeded or failed, per host.
-
-    CLI Example:
-
-    .. code-block:: bash
-
-        # Used for ESXi host connection information
-        salt '*' vsphere.set_syslog_config my.esxi.host root bad-password \
-            loghost ssl://localhost:5432,tcp://10.1.0.1:1514
-
-        # Used for connecting to a vCenter Server
-        salt '*' vsphere.set_syslog_config my.vcenter.location root bad-password \
-            loghost ssl://localhost:5432,tcp://10.1.0.1:1514 \
-            esxi_hosts='[esxi-1.host.com, esxi-2.host.com]'
-
-    """
-    ret = {}
-
-    # First, enable the syslog firewall ruleset, for each host, if needed.
-    if firewall and syslog_config == "loghost":
-        if esxi_hosts:
-            if not isinstance(esxi_hosts, list):
-                raise CommandExecutionError("'esxi_hosts' must be a list.")
-
-            for esxi_host in esxi_hosts:
-                response = enable_firewall_ruleset(
-                    host,
-                    username,
-                    password,
-                    ruleset_enable=True,
-                    ruleset_name="syslog",
-                    protocol=protocol,
-                    port=port,
-                    esxi_hosts=[esxi_host],
-                    credstore=credstore,
-                ).get(esxi_host)
-                if response["retcode"] != 0:
-                    ret.update(
-                        {
-                            esxi_host: {
-                                "enable_firewall": {
-                                    "message": response["stdout"],
-                                    "success": False,
-                                }
-                            }
-                        }
-                    )
-                else:
-                    ret.update({esxi_host: {"enable_firewall": {"success": True}}})
-        else:
-            # Handles a single host or a vCenter connection when no esxi_hosts are provided.
-            response = enable_firewall_ruleset(
-                host,
-                username,
-                password,
-                ruleset_enable=True,
-                ruleset_name="syslog",
-                protocol=protocol,
-                port=port,
-                credstore=credstore,
-            ).get(host)
-            if response["retcode"] != 0:
-                ret.update(
-                    {
-                        host: {
-                            "enable_firewall": {
-                                "message": response["stdout"],
-                                "success": False,
-                            }
-                        }
-                    }
-                )
-            else:
-                ret.update({host: {"enable_firewall": {"success": True}}})
-
-    # Set the config value on each esxi_host, if provided.
-    if esxi_hosts:
-        if not isinstance(esxi_hosts, list):
-            raise CommandExecutionError("'esxi_hosts' must be a list.")
-
-        for esxi_host in esxi_hosts:
-            response = _set_syslog_config_helper(
-                host,
-                username,
-                password,
-                syslog_config,
-                config_value,
-                protocol=protocol,
-                port=port,
-                reset_service=reset_service,
-                esxi_host=esxi_host,
-                credstore=credstore,
-            )
-            # Ensure we don't overwrite any dictionary data already set
-            # By updating the esxi_host directly.
-            if ret.get(esxi_host) is None:
-                ret.update({esxi_host: {}})
-            ret[esxi_host].update(response)
-
-    else:
-        # Handles a single host or a vCenter connection when no esxi_hosts are provided.
-        response = _set_syslog_config_helper(
-            host,
-            username,
-            password,
-            syslog_config,
-            config_value,
-            protocol=protocol,
-            port=port,
-            reset_service=reset_service,
-            credstore=credstore,
-        )
-        # Ensure we don't overwrite any dictionary data already set
-        # By updating the host directly.
-        if ret.get(host) is None:
-            ret.update({host: {}})
-        ret[host].update(response)
-
-    return ret
-
-
-@depends(HAS_ESX_CLI)
-def get_syslog_config(
-    host, username, password, protocol=None, port=None, esxi_hosts=None, credstore=None
-):
-    """
-    Retrieve the syslog configuration.
-
-    host
-        The location of the host.
-
-    username
-        The username used to login to the host, such as ``root``.
-
-    password
-        The password used to login to the host.
-
-    protocol
-        Optionally set to alternate protocol if the host is not using the default
-        protocol. Default protocol is ``https``.
-
-    port
-        Optionally set to alternate port if the host is not using the default
-        port. Default port is ``443``.
-
-    esxi_hosts
-        If ``host`` is a vCenter host, then use esxi_hosts to execute this function
-        on a list of one or more ESXi machines.
-
-    credstore
-        Optionally set to path to the credential store file.
-
-    :return: Dictionary with keys and values corresponding to the
-             syslog configuration, per host.
-
-    CLI Example:
-
-    .. code-block:: bash
-
-        # Used for ESXi host connection information
-        salt '*' vsphere.get_syslog_config my.esxi.host root bad-password
-
-        # Used for connecting to a vCenter Server
-        salt '*' vsphere.get_syslog_config my.vcenter.location root bad-password \
-            esxi_hosts='[esxi-1.host.com, esxi-2.host.com]'
-    """
-    cmd = "system syslog config get"
-
-    ret = {}
-    if esxi_hosts:
-        if not isinstance(esxi_hosts, list):
-            raise CommandExecutionError("'esxi_hosts' must be a list.")
-
-        for esxi_host in esxi_hosts:
-            response = salt.utils.vmware.esxcli(
-                host,
-                username,
-                password,
-                cmd,
-                protocol=protocol,
-                port=port,
-                esxi_host=esxi_host,
-                credstore=credstore,
-            )
-            # format the response stdout into something useful
-            ret.update({esxi_host: _format_syslog_config(response)})
-    else:
-        # Handles a single host or a vCenter connection when no esxi_hosts are provided.
-        response = salt.utils.vmware.esxcli(
-            host,
-            username,
-            password,
-            cmd,
-            protocol=protocol,
-            port=port,
-            credstore=credstore,
-        )
-        # format the response stdout into something useful
-        ret.update({host: _format_syslog_config(response)})
-
-    return ret
-
-
-@depends(HAS_ESX_CLI)
-def reset_syslog_config(
-    host,
-    username,
-    password,
-    protocol=None,
-    port=None,
-    syslog_config=None,
-    esxi_hosts=None,
-    credstore=None,
-):
-    """
-    Reset the syslog service to its default settings.
-
-    Valid syslog_config values are ``logdir``, ``loghost``, ``logdir-unique``,
-    ``default-rotate``, ``default-size``, ``default-timeout``,
-    or ``all`` for all of these.
-
-    host
-        The location of the host.
-
-    username
-        The username used to login to the host, such as ``root``.
-
-    password
-        The password used to login to the host.
-
-    protocol
-        Optionally set to alternate protocol if the host is not using the default
-        protocol. Default protocol is ``https``.
-
-    port
-        Optionally set to alternate port if the host is not using the default
-        port. Default port is ``443``.
-
-    syslog_config
-        List of parameters to reset, provided as a comma-delimited string, or 'all' to
-        reset all syslog configuration parameters. Required.
-
-    esxi_hosts
-        If ``host`` is a vCenter host, then use esxi_hosts to execute this function
-        on a list of one or more ESXi machines.
-
-    credstore
-        Optionally set to path to the credential store file.
-
-    :return: Dictionary with a top-level key of 'success' which indicates
-             if all the parameters were reset, and individual keys
-             for each parameter indicating which succeeded or failed, per host.
-
-    CLI Example:
-
-    ``syslog_config`` can be passed as a quoted, comma-separated string, e.g.
-
-    .. code-block:: bash
-
-        # Used for ESXi host connection information
-        salt '*' vsphere.reset_syslog_config my.esxi.host root bad-password \
-            syslog_config='logdir,loghost'
-
-        # Used for connecting to a vCenter Server
-        salt '*' vsphere.reset_syslog_config my.vcenter.location root bad-password \
-            syslog_config='logdir,loghost' esxi_hosts='[esxi-1.host.com, esxi-2.host.com]'
-    """
-    if not syslog_config:
-        raise CommandExecutionError(
-            "The 'reset_syslog_config' function requires a " "'syslog_config' setting."
-        )
-
-    valid_resets = [
-        "logdir",
-        "loghost",
-        "default-rotate",
-        "default-size",
-        "default-timeout",
-        "logdir-unique",
-    ]
-    cmd = "system syslog config set --reset="
-    if "," in syslog_config:
-        resets = [ind_reset.strip() for ind_reset in syslog_config.split(",")]
-    elif syslog_config == "all":
-        resets = valid_resets
-    else:
-        resets = [syslog_config]
-
-    ret = {}
-    if esxi_hosts:
-        if not isinstance(esxi_hosts, list):
-            raise CommandExecutionError("'esxi_hosts' must be a list.")
-
-        for esxi_host in esxi_hosts:
-            response_dict = _reset_syslog_config_params(
-                host,
-                username,
-                password,
-                cmd,
-                resets,
-                valid_resets,
-                protocol=protocol,
-                port=port,
-                esxi_host=esxi_host,
-                credstore=credstore,
-            )
-            ret.update({esxi_host: response_dict})
-    else:
-        # Handles a single host or a vCenter connection when no esxi_hosts are provided.
-        response_dict = _reset_syslog_config_params(
-            host,
-            username,
-            password,
-            cmd,
-            resets,
-            valid_resets,
-            protocol=protocol,
-            port=port,
-            credstore=credstore,
-        )
-        ret.update({host: response_dict})
-
-    return ret
-
-
 @ignores_kwargs("credstore")
 def upload_ssh_key(
     host,
@@ -875,290 +406,6 @@ def get_ntp_config(
 
 
 @depends(HAS_PYVMOMI)
-@ignores_kwargs("credstore")
-def get_service_policy(
-    host,
-    username,
-    password,
-    service_name,
-    protocol=None,
-    port=None,
-    host_names=None,
-    verify_ssl=True,
-):
-    """
-    Get the service name's policy for a given host or list of hosts.
-
-    host
-        The location of the host.
-
-    username
-        The username used to login to the host, such as ``root``.
-
-    password
-        The password used to login to the host.
-
-    service_name
-        The name of the service for which to retrieve the policy. Supported service names are:
-          - DCUI
-          - TSM
-          - SSH
-          - lbtd
-          - lsassd
-          - lwiod
-          - netlogond
-          - ntpd
-          - sfcbd-watchdog
-          - snmpd
-          - vprobed
-          - vpxa
-          - xorg
-
-    protocol
-        Optionally set to alternate protocol if the host is not using the default
-        protocol. Default protocol is ``https``.
-
-    port
-        Optionally set to alternate port if the host is not using the default
-        port. Default port is ``443``.
-
-    host_names
-        List of ESXi host names. When the host, username, and password credentials
-        are provided for a vCenter Server, the host_names argument is required to tell
-        vCenter the hosts for which to get service policy information.
-
-        If host_names is not provided, the service policy information will be retrieved
-        for the ``host`` location instead. This is useful for when service instance
-        connection information is used for a single ESXi host.
-
-    verify_ssl
-        Verify the SSL certificate. Default: True
-
-    CLI Example:
-
-    .. code-block:: bash
-
-        # Used for single ESXi host connection information
-        salt '*' vsphere.get_service_policy my.esxi.host root bad-password 'ssh'
-
-        # Used for connecting to a vCenter Server
-        salt '*' vsphere.get_service_policy my.vcenter.location root bad-password 'ntpd' \
-        host_names='[esxi-1.host.com, esxi-2.host.com]'
-    """
-    service_instance = salt.utils.vmware.get_service_instance(
-        host=host,
-        username=username,
-        password=password,
-        protocol=protocol,
-        port=port,
-        verify_ssl=verify_ssl,
-    )
-    valid_services = [
-        "DCUI",
-        "TSM",
-        "SSH",
-        "ssh",
-        "lbtd",
-        "lsassd",
-        "lwiod",
-        "netlogond",
-        "ntpd",
-        "sfcbd-watchdog",
-        "snmpd",
-        "vprobed",
-        "vpxa",
-        "xorg",
-    ]
-    host_names = _check_hosts(service_instance, host, host_names)
-
-    ret = {}
-    for host_name in host_names:
-        # Check if the service_name provided is a valid one.
-        # If we don't have a valid service, return. The service will be invalid for all hosts.
-        if service_name not in valid_services:
-            ret.update(
-                {
-                    host_name: {
-                        "Error": "{} is not a valid service name.".format(service_name)
-                    }
-                }
-            )
-            return ret
-
-        host_ref = _get_host_ref(service_instance, host, host_name=host_name)
-        services = host_ref.configManager.serviceSystem.serviceInfo.service
-
-        # Don't require users to know that VMware lists the ssh service as TSM-SSH
-        if service_name == "SSH" or service_name == "ssh":
-            temp_service_name = "TSM-SSH"
-        else:
-            temp_service_name = service_name
-
-        # Loop through services until we find a matching name
-        for service in services:
-            if service.key == temp_service_name:
-                ret.update({host_name: {service_name: service.policy}})
-                # We've found a match - break out of the loop so we don't overwrite the
-                # Updated host_name value with an error message.
-                break
-            else:
-                msg = "Could not find service '{}' for host '{}'.".format(
-                    service_name, host_name
-                )
-                ret.update({host_name: {"Error": msg}})
-
-        # If we made it this far, something else has gone wrong.
-        if ret.get(host_name) is None:
-            msg = "'vsphere.get_service_policy' failed for host {}.".format(host_name)
-            log.debug(msg)
-            ret.update({host_name: {"Error": msg}})
-
-    return ret
-
-
-@depends(HAS_PYVMOMI)
-@ignores_kwargs("credstore")
-def get_service_running(
-    host,
-    username,
-    password,
-    service_name,
-    protocol=None,
-    port=None,
-    host_names=None,
-    verify_ssl=True,
-):
-    """
-    Get the service name's running state for a given host or list of hosts.
-
-    host
-        The location of the host.
-
-    username
-        The username used to login to the host, such as ``root``.
-
-    password
-        The password used to login to the host.
-
-    service_name
-        The name of the service for which to retrieve the policy. Supported service names are:
-          - DCUI
-          - TSM
-          - SSH
-          - lbtd
-          - lsassd
-          - lwiod
-          - netlogond
-          - ntpd
-          - sfcbd-watchdog
-          - snmpd
-          - vprobed
-          - vpxa
-          - xorg
-
-    protocol
-        Optionally set to alternate protocol if the host is not using the default
-        protocol. Default protocol is ``https``.
-
-    port
-        Optionally set to alternate port if the host is not using the default
-        port. Default port is ``443``.
-
-    host_names
-        List of ESXi host names. When the host, username, and password credentials
-        are provided for a vCenter Server, the host_names argument is required to tell
-        vCenter the hosts for which to get the service's running state.
-
-        If host_names is not provided, the service's running state will be retrieved
-        for the ``host`` location instead. This is useful for when service instance
-        connection information is used for a single ESXi host.
-
-    verify_ssl
-        Verify the SSL certificate. Default: True
-
-    CLI Example:
-
-    .. code-block:: bash
-
-        # Used for single ESXi host connection information
-        salt '*' vsphere.get_service_running my.esxi.host root bad-password 'ssh'
-
-        # Used for connecting to a vCenter Server
-        salt '*' vsphere.get_service_running my.vcenter.location root bad-password 'ntpd' \
-        host_names='[esxi-1.host.com, esxi-2.host.com]'
-    """
-    service_instance = salt.utils.vmware.get_service_instance(
-        host=host,
-        username=username,
-        password=password,
-        protocol=protocol,
-        port=port,
-        verify_ssl=verify_ssl,
-    )
-    valid_services = [
-        "DCUI",
-        "TSM",
-        "SSH",
-        "ssh",
-        "lbtd",
-        "lsassd",
-        "lwiod",
-        "netlogond",
-        "ntpd",
-        "sfcbd-watchdog",
-        "snmpd",
-        "vprobed",
-        "vpxa",
-        "xorg",
-    ]
-    host_names = _check_hosts(service_instance, host, host_names)
-
-    ret = {}
-    for host_name in host_names:
-        # Check if the service_name provided is a valid one.
-        # If we don't have a valid service, return. The service will be invalid for all hosts.
-        if service_name not in valid_services:
-            ret.update(
-                {
-                    host_name: {
-                        "Error": "{} is not a valid service name.".format(service_name)
-                    }
-                }
-            )
-            return ret
-
-        host_ref = _get_host_ref(service_instance, host, host_name=host_name)
-        services = host_ref.configManager.serviceSystem.serviceInfo.service
-
-        # Don't require users to know that VMware lists the ssh service as TSM-SSH
-        if service_name == "SSH" or service_name == "ssh":
-            temp_service_name = "TSM-SSH"
-        else:
-            temp_service_name = service_name
-
-        # Loop through services until we find a matching name
-        for service in services:
-            if service.key == temp_service_name:
-                ret.update({host_name: {service_name: service.running}})
-                # We've found a match - break out of the loop so we don't overwrite the
-                # Updated host_name value with an error message.
-                break
-            else:
-                msg = "Could not find service '{}' for host '{}'.".format(
-                    service_name, host_name
-                )
-                ret.update({host_name: {"Error": msg}})
-
-        # If we made it this far, something else has gone wrong.
-        if ret.get(host_name) is None:
-            msg = "'vsphere.get_service_running' failed for host {}.".format(host_name)
-            log.debug(msg)
-            ret.update({host_name: {"Error": msg}})
-
-    return ret
-
-
-@depends(HAS_PYVMOMI)
 @_supports_proxies("esxi", "esxcluster", "esxdatacenter", "vcenter", "esxvm")
 @_gets_service_instance_via_proxy
 def test_vcenter_connection(service_instance=None):
@@ -1230,95 +477,6 @@ def system_info(
     return ret
 
 
-
-@depends(HAS_PYVMOMI)
-@ignores_kwargs("credstore")
-def list_datastore_clusters(
-    host, username, password, protocol=None, port=None, verify_ssl=True
-):
-    """
-    Returns a list of datastore clusters for the specified host.
-
-    host
-        The location of the host.
-
-    username
-        The username used to login to the host, such as ``root``.
-
-    password
-        The password used to login to the host.
-
-    protocol
-        Optionally set to alternate protocol if the host is not using the default
-        protocol. Default protocol is ``https``.
-
-    port
-        Optionally set to alternate port if the host is not using the default
-        port. Default port is ``443``.
-
-    verify_ssl
-        Verify the SSL certificate. Default: True
-
-    CLI Example:
-
-    .. code-block:: bash
-
-        salt '*' vsphere.list_datastore_clusters 1.2.3.4 root bad-password
-    """
-    service_instance = salt.utils.vmware.get_service_instance(
-        host=host,
-        username=username,
-        password=password,
-        protocol=protocol,
-        port=port,
-        verify_ssl=verify_ssl,
-    )
-    return salt.utils.vmware.list_datastore_clusters(service_instance)
-
-
-@depends(HAS_PYVMOMI)
-@ignores_kwargs("credstore")
-def list_datastores(
-    host, username, password, protocol=None, port=None, verify_ssl=True
-):
-    """
-    Returns a list of datastores for the specified host.
-
-    host
-        The location of the host.
-
-    username
-        The username used to login to the host, such as ``root``.
-
-    password
-        The password used to login to the host.
-
-    protocol
-        Optionally set to alternate protocol if the host is not using the default
-        protocol. Default protocol is ``https``.
-
-    port
-        Optionally set to alternate port if the host is not using the default
-        port. Default port is ``443``.
-
-    verify_ssl
-        Verify the SSL certificate. Default: True
-
-    CLI Example:
-
-    .. code-block:: bash
-
-        salt '*' vsphere.list_datastores 1.2.3.4 root bad-password
-    """
-    service_instance = salt.utils.vmware.get_service_instance(
-        host=host,
-        username=username,
-        password=password,
-        protocol=protocol,
-        port=port,
-        verify_ssl=verify_ssl,
-    )
-    return salt.utils.vmware.list_datastores(service_instance)
 
 
 @depends(HAS_PYVMOMI)
@@ -2163,135 +1321,6 @@ def update_storage_policy(policy, policy_dict, service_instance=None):
     return {"update_storage_policy": True}
 
 
-@depends(HAS_PYVMOMI)
-@_supports_proxies("esxcluster", "esxdatacenter", "vcenter")
-@_gets_service_instance_via_proxy
-def list_default_storage_policy_of_datastore(datastore, service_instance=None):
-    """
-    Returns a list of datastores assign the storage policies.
-
-    datastore
-        Name of the datastore to assign.
-        The datastore needs to be visible to the VMware entity the proxy
-        points to.
-
-    service_instance
-        Service instance (vim.ServiceInstance) of the vCenter.
-        Default is None.
-
-    .. code-block:: bash
-
-        salt '*' vsphere.list_default_storage_policy_of_datastore datastore=ds1
-    """
-    log.trace(
-        "Listing the default storage policy of datastore '{}'" "".format(datastore)
-    )
-    # Find datastore
-    target_ref = _get_proxy_target(service_instance)
-    ds_refs = salt.utils.vmware.get_datastores(
-        service_instance, target_ref, datastore_names=[datastore]
-    )
-    if not ds_refs:
-        raise VMwareObjectRetrievalError(
-            "Datastore '{}' was not " "found".format(datastore)
-        )
-    profile_manager = salt.utils.pbm.get_profile_manager(service_instance)
-    policy = salt.utils.pbm.get_default_storage_policy_of_datastore(
-        profile_manager, ds_refs[0]
-    )
-    return _get_policy_dict(policy)
-
-
-@depends(HAS_PYVMOMI)
-@_supports_proxies("esxcluster", "esxdatacenter", "vcenter")
-@_gets_service_instance_via_proxy
-def assign_default_storage_policy_to_datastore(
-    policy, datastore, service_instance=None
-):
-    """
-    Assigns a storage policy as the default policy to a datastore.
-
-    policy
-        Name of the policy to assign.
-
-    datastore
-        Name of the datastore to assign.
-        The datastore needs to be visible to the VMware entity the proxy
-        points to.
-
-    service_instance
-        Service instance (vim.ServiceInstance) of the vCenter.
-        Default is None.
-
-    .. code-block:: bash
-
-        salt '*' vsphere.assign_storage_policy_to_datastore
-            policy='policy name' datastore=ds1
-    """
-    log.trace("Assigning policy {} to datastore {}" "".format(policy, datastore))
-    profile_manager = salt.utils.pbm.get_profile_manager(service_instance)
-    # Find policy
-    policies = salt.utils.pbm.get_storage_policies(profile_manager, [policy])
-    if not policies:
-        raise VMwareObjectRetrievalError("Policy '{}' was not found" "".format(policy))
-    policy_ref = policies[0]
-    # Find datastore
-    target_ref = _get_proxy_target(service_instance)
-    ds_refs = salt.utils.vmware.get_datastores(
-        service_instance, target_ref, datastore_names=[datastore]
-    )
-    if not ds_refs:
-        raise VMwareObjectRetrievalError(
-            "Datastore '{}' was not " "found".format(datastore)
-        )
-    ds_ref = ds_refs[0]
-    salt.utils.pbm.assign_default_storage_policy_to_datastore(
-        profile_manager, policy_ref, ds_ref
-    )
-    return True
-
-
-@depends(HAS_PYVMOMI)
-@_supports_proxies("esxdatacenter", "esxcluster", "vcenter", "esxvm")
-@_gets_service_instance_via_proxy
-def list_datacenters_via_proxy(datacenter_names=None, service_instance=None):
-    """
-    Returns a list of dict representations of VMware datacenters.
-    Connection is done via the proxy details.
-
-    Supported proxies: esxdatacenter
-
-    datacenter_names
-        List of datacenter names.
-        Default is None.
-
-    service_instance
-        Service instance (vim.ServiceInstance) of the vCenter.
-        Default is None.
-
-    .. code-block:: bash
-
-        salt '*' vsphere.list_datacenters_via_proxy
-
-        salt '*' vsphere.list_datacenters_via_proxy dc1
-
-        salt '*' vsphere.list_datacenters_via_proxy dc1,dc2
-
-        salt '*' vsphere.list_datacenters_via_proxy datacenter_names=[dc1, dc2]
-    """
-    if not datacenter_names:
-        dc_refs = salt.utils.vmware.get_datacenters(
-            service_instance, get_all_datacenters=True
-        )
-    else:
-        dc_refs = salt.utils.vmware.get_datacenters(service_instance, datacenter_names)
-
-    return [
-        {"name": salt.utils.vmware.get_managed_object_name(dc_ref)}
-        for dc_ref in dc_refs
-    ]
-
-
 def _get_cluster_dict(cluster_name, cluster_ref):
     """
     Returns a cluster dict representation from
@@ -2508,244 +1537,6 @@ def _apply_cluster_dict(cluster_spec, cluster_dict, vsan_spec=None, vsan_61=True
     log.trace("cluster_spec = {}".format(cluster_spec))
 
 
-@depends(HAS_PYVMOMI)
-@_supports_proxies("esxi", "esxcluster", "esxdatacenter")
-@_gets_service_instance_via_proxy
-def list_datastores_via_proxy(
-    datastore_names=None,
-    backing_disk_ids=None,
-    backing_disk_scsi_addresses=None,
-    service_instance=None,
-):
-    """
-    Returns a list of dict representations of the datastores visible to the
-    proxy object. The list of datastores can be filtered by datastore names,
-    backing disk ids (canonical names) or backing disk scsi addresses.
-
-    Supported proxy types: esxi, esxcluster, esxdatacenter
-
-    datastore_names
-        List of the names of datastores to filter on
-
-    backing_disk_ids
-        List of canonical names of the backing disks of the datastores to filer.
-        Default is None.
-
-    backing_disk_scsi_addresses
-        List of scsi addresses of the backing disks of the datastores to filter.
-        Default is None.
-
-    service_instance
-        Service instance (vim.ServiceInstance) of the vCenter/ESXi host.
-        Default is None.
-
-    .. code-block:: bash
-
-        salt '*' vsphere.list_datastores_via_proxy
-
-        salt '*' vsphere.list_datastores_via_proxy datastore_names=[ds1, ds2]
-    """
-    target = _get_proxy_target(service_instance)
-    target_name = salt.utils.vmware.get_managed_object_name(target)
-    log.trace("target name = {}".format(target_name))
-
-    # Default to getting all disks if no filtering is done
-    get_all_datastores = (
-        True
-        if not (datastore_names or backing_disk_ids or backing_disk_scsi_addresses)
-        else False
-    )
-    # Get the ids of the disks with the scsi addresses
-    if backing_disk_scsi_addresses:
-        log.debug(
-            "Retrieving disk ids for scsi addresses "
-            "'{}'".format(backing_disk_scsi_addresses)
-        )
-        disk_ids = [
-            d.canonicalName
-            for d in salt.utils.vmware.get_disks(
-                target, scsi_addresses=backing_disk_scsi_addresses
-            )
-        ]
-        log.debug("Found disk ids '{}'".format(disk_ids))
-        backing_disk_ids = (
-            backing_disk_ids.extend(disk_ids) if backing_disk_ids else disk_ids
-        )
-    datastores = salt.utils.vmware.get_datastores(
-        service_instance, target, datastore_names, backing_disk_ids, get_all_datastores
-    )
-
-    # Search for disk backed datastores if target is host
-    # to be able to add the backing_disk_ids
-    mount_infos = []
-    if isinstance(target, vim.HostSystem):
-        storage_system = salt.utils.vmware.get_storage_system(
-            service_instance, target, target_name
-        )
-        props = salt.utils.vmware.get_properties_of_managed_object(
-            storage_system, ["fileSystemVolumeInfo.mountInfo"]
-        )
-        mount_infos = props.get("fileSystemVolumeInfo.mountInfo", [])
-    ret_dict = []
-    for ds in datastores:
-        ds_dict = {
-            "name": ds.name,
-            "type": ds.summary.type,
-            "free_space": ds.summary.freeSpace,
-            "capacity": ds.summary.capacity,
-        }
-        backing_disk_ids = []
-        for vol in [
-            i.volume
-            for i in mount_infos
-            if i.volume.name == ds.name and isinstance(i.volume, vim.HostVmfsVolume)
-        ]:
-
-            backing_disk_ids.extend([e.diskName for e in vol.extent])
-        if backing_disk_ids:
-            ds_dict["backing_disk_ids"] = backing_disk_ids
-        ret_dict.append(ds_dict)
-    return ret_dict
-
-
-@depends(HAS_PYVMOMI)
-@depends(HAS_JSONSCHEMA)
-@_supports_proxies("esxi")
-@_gets_service_instance_via_proxy
-def create_vmfs_datastore(
-    datastore_name,
-    disk_id,
-    vmfs_major_version,
-    safety_checks=True,
-    service_instance=None,
-):
-    """
-    Creates a ESXi host disk group with the specified cache and capacity disks.
-
-    datastore_name
-        The name of the datastore to be created.
-
-    disk_id
-        The disk id (canonical name) on which the datastore is created.
-
-    vmfs_major_version
-        The VMFS major version.
-
-    safety_checks
-        Specify whether to perform safety check or to skip the checks and try
-        performing the required task. Default is True.
-
-    service_instance
-        Service instance (vim.ServiceInstance) of the vCenter/ESXi host.
-        Default is None.
-
-    .. code-block:: bash
-
-        salt '*' vsphere.create_vmfs_datastore datastore_name=ds1 disk_id=
-            vmfs_major_version=5
-    """
-    log.debug("Validating vmfs datastore input")
-    schema = VmfsDatastoreSchema.serialize()
-    try:
-        jsonschema.validate(
-            {
-                "datastore": {
-                    "name": datastore_name,
-                    "backing_disk_id": disk_id,
-                    "vmfs_version": vmfs_major_version,
-                }
-            },
-            schema,
-        )
-    except jsonschema.exceptions.ValidationError as exc:
-        raise ArgumentValueError(exc)
-    host_ref = _get_proxy_target(service_instance)
-    hostname = __proxy__["esxi.get_details"]()["esxi_host"]
-    if safety_checks:
-        disks = salt.utils.vmware.get_disks(host_ref, disk_ids=[disk_id])
-        if not disks:
-            raise VMwareObjectRetrievalError(
-                "Disk '{}' was not found in host '{}'".format(disk_id, hostname)
-            )
-    ds_ref = salt.utils.vmware.create_vmfs_datastore(
-        host_ref, datastore_name, disks[0], vmfs_major_version
-    )
-    return True
-
-
-@depends(HAS_PYVMOMI)
-@_supports_proxies("esxi", "esxcluster", "esxdatacenter")
-@_gets_service_instance_via_proxy
-def rename_datastore(datastore_name, new_datastore_name, service_instance=None):
-    """
-    Renames a datastore. The datastore needs to be visible to the proxy.
-
-    datastore_name
-        Current datastore name.
-
-    new_datastore_name
-        New datastore name.
-
-    service_instance
-        Service instance (vim.ServiceInstance) of the vCenter/ESXi host.
-        Default is None.
-
-    .. code-block:: bash
-
-        salt '*' vsphere.rename_datastore old_name new_name
-    """
-    # Argument validation
-    log.trace(
-        "Renaming datastore {} to {}" "".format(datastore_name, new_datastore_name)
-    )
-    target = _get_proxy_target(service_instance)
-    datastores = salt.utils.vmware.get_datastores(
-        service_instance, target, datastore_names=[datastore_name]
-    )
-    if not datastores:
-        raise VMwareObjectRetrievalError(
-            "Datastore '{}' was not found" "".format(datastore_name)
-        )
-    ds = datastores[0]
-    salt.utils.vmware.rename_datastore(ds, new_datastore_name)
-    return True
-
-
-@depends(HAS_PYVMOMI)
-@_supports_proxies("esxi", "esxcluster", "esxdatacenter")
-@_gets_service_instance_via_proxy
-def remove_datastore(datastore, service_instance=None):
-    """
-    Removes a datastore. If multiple datastores an error is raised.
-
-    datastore
-        Datastore name
-
-    service_instance
-        Service instance (vim.ServiceInstance) of the vCenter/ESXi host.
-        Default is None.
-
-    .. code-block:: bash
-
-        salt '*' vsphere.remove_datastore ds_name
-    """
-    log.trace("Removing datastore '{}'".format(datastore))
-    target = _get_proxy_target(service_instance)
-    datastores = salt.utils.vmware.get_datastores(
-        service_instance, reference=target, datastore_names=[datastore]
-    )
-    if not datastores:
-        raise VMwareObjectRetrievalError(
-            "Datastore '{}' was not found".format(datastore)
-        )
-    if len(datastores) > 1:
-        raise VMwareObjectRetrievalError(
-            "Multiple datastores '{}' were found".format(datastore)
-        )
-    salt.utils.vmware.remove_datastore(service_instance, datastores[0])
-    return True
-
-
 def _get_entity(service_instance, entity):
     """
     Returns the entity associated with the entity dict representation
@@ -2948,77 +1739,6 @@ def _check_hosts(service_instance, host, host_names):
         raise CommandExecutionError("'host_names' must be a list.")
 
     return host_names
-
-
-def _format_coredump_stdout(cmd_ret):
-    """
-    Helper function to format the stdout from the get_coredump_network_config function.
-
-    cmd_ret
-        The return dictionary that comes from a cmd.run_all call.
-    """
-    ret_dict = {}
-    for line in cmd_ret["stdout"].splitlines():
-        line = line.strip().lower()
-        if line.startswith("enabled:"):
-            enabled = line.split(":")
-            if "true" in enabled[1]:
-                ret_dict["enabled"] = True
-            else:
-                ret_dict["enabled"] = False
-                break
-        if line.startswith("host vnic:"):
-            host_vnic = line.split(":")
-            ret_dict["host_vnic"] = host_vnic[1].strip()
-        if line.startswith("network server ip:"):
-            ip = line.split(":")
-            ret_dict["ip"] = ip[1].strip()
-        if line.startswith("network server port:"):
-            ip_port = line.split(":")
-            ret_dict["port"] = ip_port[1].strip()
-
-    return ret_dict
-
-
-def _format_firewall_stdout(cmd_ret):
-    """
-    Helper function to format the stdout from the get_firewall_status function.
-
-    cmd_ret
-        The return dictionary that comes from a cmd.run_all call.
-    """
-    ret_dict = {"success": True, "rulesets": {}}
-    for line in cmd_ret["stdout"].splitlines():
-        if line.startswith("Name"):
-            continue
-        if line.startswith("---"):
-            continue
-        ruleset_status = line.split()
-        ret_dict["rulesets"][ruleset_status[0]] = bool(ruleset_status[1])
-
-    return ret_dict
-
-
-def _format_syslog_config(cmd_ret):
-    """
-    Helper function to format the stdout from the get_syslog_config function.
-
-    cmd_ret
-        The return dictionary that comes from a cmd.run_all call.
-    """
-    ret_dict = {"success": cmd_ret["retcode"] == 0}
-
-    if cmd_ret["retcode"] != 0:
-        ret_dict["message"] = cmd_ret["stdout"]
-    else:
-        for line in cmd_ret["stdout"].splitlines():
-            line = line.strip()
-            cfgvars = line.split(": ")
-            key = cfgvars[0].strip()
-            value = cfgvars[1].strip()
-            ret_dict[key] = value
-
-    return ret_dict
 
 
 def _get_date_time_mgr(host_reference):
